more cleanup

+++ src/main/java/com/florian/videoclient/client/activities/videolist/Video.java
@@ -9,7 +9,7 @@ public class Video extends JavaScriptObject {
     public final native String getId() /*-{ return this.id; }-*/;
     public final native String getName() /*-{ return this.name; }-*/;
     public final native String getUrl() /*-{ return this.url; }-*/;
-    public final native Long getDuration() /*-{ return this.duration; }-*/;
+    public final native int getDuration() /*-{ return this.duration; }-*/;
 
     public final native void setName(String name) /*-{this.name = name; }-*/;
     public final native void setUrl(String url) /*-{this.url = url; }-*/;
